Handle persistent mode differently.
Earlier persistence mode was not set on GPU's, which healthagent
attempted to do at the start. This fails as now nvidia-persistenced
service is constantly running. Now we should track persistence mode
and show an error if the service is not running. We track it both
by DCGM fields as well as through systemd monitoring.