Add a test mode for DCGM
This allows healthagent GPU tests to run via nvidia-dcgm service.
This is not the default mode. By default DCGM library is loaded in
embedded mode to avoid reliance on nvidia-dcgm service in production.

In test-mode healthagent loads the dcgm library by connecting to
nvidia-dcgm service and this facilitates testing of DCGM api.
This mode is only for testing.